refactor(webhooks): isolate request schemas and trim duplicate plumbing
